[FIRRTL] Use walk, InstanceInfo in AddSeqMemPorts
Change the `AddSeqMemPorts` pass (whic adds MBIST ports to memories) to
use a walk when adding ports to memories.  Practically, this causes the
pass to now be capable of adding MBIST to memories under `WhenOp`s or
`LayerBlockOp`s.  However, because this pass is intentionally _not_
supposed to add MBIST ports to anything that is not in the "design", this
will skip over any layers or modules that are instantiated wholly under
layers.

To do this, this pass was changed to use the new `InstanceInfo` analysis
instead of computing information about what modules are under the
design-under-test (DUT) manually.  Any checking related to layers is
similarly done using this analysis.

This change is prerequisite work to move the `LowerLayers` after
`AddSeqMemPorts` while still preserving the functionality of the original
pass order.  Unfortunately, layers do not compose sanely with
`AddSeqMemPorts`.  This is a failing of the original pass and its
assumptions about what is "design" vs. "verification" as opposed to any
fundamental issue with layers.  Generally, layers are intended as
verification features.  However, there is no actual requirement or
expectation that they will be used this way.  Going forward,
`AddSeqMemPorts` is intended to be removed and replaced with a feature in
Chisel.

Additionally, this this commit includes a small change to the pass output
where the connects created are placed immediately following the memory
instance.  This is necessary to create the connects inside the layer
blocks.  However, it is also slightly better output than before.  (This
changed one test superficially.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>