[FIRRTL] Switch to IPO instead of RPO
Change FIRRTL's `InstanceInfo` analysis to use `llvm::inverse_post_order`
instead of `ReversePostOrderTraversal`.  This may be slightly faster/more
efficient as it doesn't build a vector of what to visit in a post order
walk and this will visit all modules, not just those under the top module.

h/t @youngar for the pattern and the reasoning behind why to favor this
pattern.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>