[FIRRTL] Add verifier of single MarkDUTAnnotation (#7633)
Add a CircuitOp verifier that guarantees that there is zero or one modules
annotated with a `MarkDUTAnnotation`.  This is important as many passes
assume that this is the case.  Moving this check into a verifier means
that logic checking this behavior in each pass can be removed.  I am
slightly concerned of the performance cost of doing this as examining the
annotations is non-trivial and this will cause that examination to now
happen more often.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>