squash! [FIRRTL] Use walk, InstanceInfo in AddSeqMemPorts
Make AddSeqMemPorts work for memories under WhenOps.  This takes a necessary
interpretation of how this should work which may create disagreement.  However,
it is the only reasonable thing to do.  Namely, for a memory that is under a
WhenOp that is having _output_ ports added to it, this will drive the port under
the WhenOp and invalidate it at the top of the module.

This interpretation is that the MBIST outputs are only driven under the WhenOp
and indeterminate otherwise.  CIRCT will treat this as an unconditional connect
after ExpandWhens, however, the spec allows for it to connect it to anything.