[FIRRTL] Use walk, InstanceInfo in AddSeqMemPorts (#7599)
Change the `AddSeqMemPorts` pass (which adds MBIST ports to memories) to
use a walk when adding ports to memories.  This will visit any region, but
will skip memories which instantiated under LayerBlockOps.  Memories and
instances which are instantiated under WhenOps will now get ports added to
them.  Any output ports created for a memory under a WhenOp will be
invalidated.

This change is prerequisite work to move the `LowerLayers` after
`AddSeqMemPorts` while still preserving the functionality of the original
pass order.  Unfortunately, layers do not compose sanely with
`AddSeqMemPorts`.  This is a failing of the original pass and its
assumptions about what is "design" vs. "verification" as opposed to any
fundamental issue with layers.  Generally, layers are intended as
verification features.  However, there is no actual requirement or
expectation that they will be used this way.  Going forward,
`AddSeqMemPorts` is intended to be removed and replaced with a feature in
Chisel.

Additionally, this this commit includes a small change to the pass output
where the connects created are placed immediately following the memory
instance.  This is necessary to create the connects inside the layer
blocks.  However, it is also slightly better output than before.  (This
changed one test superficially.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>