[FIRRTL] Error mixed DUT mems in AddSeqMemPorts (#7622)
Change AddSeqMemPorts to reject circuits where a memory is instantiated
under both the design-under-test (DUT) and _not_ under the DUT.  This is a
situation that indicates that ports should be added to one instance, but
not the other.  While there are a number of ways that this could be
solved (similarly to with mixed under layer / _not_ under layer) by
transforming the design via recursive deduplication or by tying off the
modified memory, these approaches are rejected for now in favor of just
catching the error.  Mid-term I would prefer to delete this pass entirely
as it has general composition problems with other passes and FIRRTL
features.

Fixes #7620.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>