[FIRRTL][LowerLayers] Remove handling of some ref ops (#7640)
Remove code that handles the capturing of ref cast and ref send operands specially, since
this code is doing only what the generic capture-handling code does just a few lines below.

There is a bug in the handling of ref.cast which is removed by this change: We
are accidentally checking that the cast op itself is within the layerblock,
when we should actually be checking if its operand is defined within the
layerblock. This causes us to incorrectly capture the operand of a ref.cast op.
The generic codepath works fine.