[FIRRTL] InferWidths: simplify allWidthsKnownCheck
When all result widths of an operation are known, we do not have to
create a constraint expression based on the operands, we can directly
create a known width expression.  The check for known widths was a bit
convoluted, checking some things in a loop which did not need to be.
This change moves the functionality into a helper method for easier to
understand control flow, and moves the mux operand check outside of a
loop.