[FIRRTL] InferWidths: reuse workstack while solving
We would allocate a new workstack each time we solve a var, and reserve
a size of 1/2 of the total number of expressions.  The number of
expressions in a ciruit is absolutely massive, making this not a good
default.  Instead, we just allocate the workstack from outside of the
method and reuse the memory allocation for each solve.

As an aside, this fixes the indentation for debug printing which was
broken in #5305.