[Verif] Simplify the FormalOp
Simplify the definition of the `verif.formal` operation which describes
a single formal unit test. This op used to have a "k" parameter which is
now gone in favor of more generically using the attributes dictionary to
capture execution parameters. As we build out the testing flows, we can
reintroduce necessary attributes to this operation.

Rename `verif.symbolic_input` to `verif.symbolic_value`.

Remove `verif.concrete_input` in favor of explicitly describing the
concrete input using one of the registers in the Seq dialect. This also
makes the associated clock explicit, which was missing from this op.