[FIRRTL] InferWidths: Don't need to duplicate InvalidValues
We used to duplicate InvalidValues to their uses, so that each use could
be inferred separately.  This was because InvalidValues were considered
constants, and things like CSE might combine many unknown-width invalid
values, and we wanted to counteract this affect.  InvalidValues are no
longer pure, and so will not be CSE'd away this way anymore, making this
code no longer needed.