[ExtractInstances] Add the extract instances metadata to OM Classes (#7667)
This commit adds the ExtractInstances metadata to the OM classes. The emitted
OM dialect can then be parsed to generate the respective metadata files. This
change enables to move all the metadata to the OM dialect, instead of existing
as verbatim ops.
Also move some common ClassOp utilities to the builder and update the
CreateSifiveMetadata pass accordingly.