[FIRRTL] Don't force non-local trackers in Dedup.
The previous behavior for local trackers would be to clone the tracker
into duplicate copies for each unique instance. This will fail in
LowerClasses, because one path op that used to refer to a single
entity as a local reference would now refer to every hierarchical
instance of that entity. One path can't refer to multiple things, so
if the user specified a local reference, allow that to pass through
without expanding it into every hierarchical instance.