[Verif][NFC] Used auto-generated constructors for all passes
Drop the `let constructor` line from the Verif pass definitions. This
causes TableGen to automatically generate the constructors for these
passes, and also handle pass options if we ever decide to add them.

This requires adding a `...Pass` to the end of each pass def to produce
a constructor name of the form `create...Pass()`.