[FIRRTL] Make memory matadata work with layers
Extend the existing notion of how memory metadata is generated to work for
memories that are under layers.  Use an interpretation of the "design"
excluding layers.  In effect, memories which are under layers now generate
metadata as if that instantiation was in the test harness.

I note that the current behavior here is inconsistent.  The metadata
generation behavior is different for a memory which is partially in the
design (it has instances which are in the design and not in the design)
compared to a memory which is wholly not in the design.  However, this is
the exact same behavior as before this change.

The details of what happens here are somewhat complex.  For documentation
of the current behavior, see the tests that accompany this commit.

In the future, this behavior should be changed to be dramatically simpler.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>