[FIRRTL] ReplSeqMems: use the original mem name for the instance name (#7776)
When LowerMemories is replacing sequential memories with instances of
extmodules, it uses the name of the of the generated extmodule as the
name of the instance. This changes the pass to instead use the to use
the original memory name + "_ext" for the instance of the extmodule.
This mostly effects the instance name when a single extmodule was
generated for two different memories with different names.