[FIRRTL] Cleanup CreateSiFiveMetadata tests, NFC
Make additional modifications to the `emit-metadata.mlir` test to test
more features orthogonally.  This includes a test of a zero-port
memory (which apparently is not rejected by verification?) and a separate
read-only memory test.  This replaces a two-read memory (which is already
covered with the original omnibus test) and the older read-only memory
test.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>