[ESI][Runtime] Fix wheel builds (#7795)
CMake 3.31 changed the behavior of `install(IMPORTED_RUNTIME_ARTIFACTS` to actually install them instead of merely adding them to the list. This ended up pulling in _all_ of the shared libs (including system libraries like `libm`, `libgcc`, etc.) instead of respecting the filters given in `install(RUNTIME_DEPENDENCY_SET`. As a result, auditwheel was failing. Happened over the weekend when a new version of the `cmake` package was published to PyPI.