[FIRRTL] Run canonicalizer again after IMCP
IMCP propagates constants across modules, which should enable more
canonicalization.  However, lower-layers, which runs between IMCP and
canonicalization, takes layerblocks and outlines them into their own module,
which effectively undoes the work of IMCP, and prevents us from canonicalizing
as well as we should under layers. This PR runs the canonicalizer immediately
after IMCP, so that we have a better shot at canonicalization.

In particular, we want to ensure that we are able to canonicalize away dead
assertions and assumptions under layerblocks.