[FIRRTL] Add layer support to ExtractInstances
Update the `ExtractInstances` pass to not extract modules which are
instantiated under layers.  For situations where a module is both in the
design and under a layer, the module will be extracted up to the point
where the instance is outside the design.  This is the same behavior that
the pass does if a module is instantiated by both the test harness and the
design.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>