[FIRRTL] Bump nextFIRVersion to 4.0.0 in the parser
Mark all parser features that were part of FIRRTL spec 3.3.0 [1] as
requiring version `{3, 3, 0}` instead of `nextFIRVersion`. Bump the
`nextFIRVersion` to be 4.0.0, matching the `exportFIRVersion` as it used
to be the case earlier. Note that specifically the "List concat" feature
was not part of 3.3.0 and is therefore still marked as `nextFIRVersion`.

Also add a `missingSpecFIRVersion` to mark the features in the parser
that have not been added to the spec yet. This avoids accidental
replacement of the version during the next spec release.

This change requires a few tweaks to the regression tests, where a few
features like `class`, `Bool`, `Double`, `Inst`, `path`, and more, were
tested under FIRRTL 3.3.0, assuming they would be released as part of
the spec in that version. But the features have never made their way
into the spec, so all of these tests now have to point at the next spec
version.

[1]: https://github.com/chipsalliance/firrtl-spec/releases/tag/v3.3.0