[FIRRTL] Update, fix parser versions
FIRRTL v4.0.0 was released [[1]].  Update the FIRRTL parser to correctly
accept/reject features from this version now that the feature has
solidified.  Fix a number of incorrect version checks.

Note: when going through this, I realized that the 3.3.0 version made some
breaking changes, notably with respect to "optional groups".  These were
added in 3.2.0 and dropped in 3.3.0 (in favor of layers).  This isn't
hugely problematic, however, it did violate the FIRRTL specification
versioning policy.  Oops.

The proliferation of version checks on FIRRTL features is really starting
to make me think that we want to put this into a tabular structure
indicating: (1) when a feature was added and (2) when a feature was
removed.  We can then just check all these features automatically as
opposed to sprinkling the checks around.

[1]: https://github.com/chipsalliance/firrtl-spec/releases/tag/v4.0.0

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>