[FIRRTL] Lower firrtl.formal to verif.formal
Add a lowering from `firrtl.formal` to `verif.formal` in the
FIRRTL-to-HW conversion. The lowering creates a `verif.formal` op that
simply instantiates the module pointed to by `firrtl.formal`, with
symbolic values providing the inputs to the module. We may want to
inline the instance in the future, but that is purely cosmetic.

This now allows FIR files to contain `formal` unti tests that can be
lowered to HW through `firtool` and then executed using `circt-test`.