Update calls to APInt constructor.
This was needed when
https://github.com/llvm/llvm-project/commit/3494ee95902cef62f767489802e469c58a13ea04
landed.

Similar changes were made upstream here:
https://github.com/llvm/llvm-project/pull/80309/files.

Basically, we now expect the width and value arguments to the APInt
constructor to be consistent. There is some implicit extension or
truncation depending on the signedness of the value, and now a flag to
opt into it.

With this change I've simply opted into the flag where necessary,
rather than thinking hard about how to compute the correct width given
the value.

In places where the `isSigned` flag took the default value of false, I
kept that when I had to explicitly set that flag.

There was exactly one place (CalyxHelpers.cpp), where I actually
flipped the value of `isSigned`. The comment in the code made me think
this erroneously thought the flag was "isUnsigned".