[FIRRTL][docs] GC Views are non-design
Update language in the FIRRTL Rationale to indicate that Grand Central
Views are not part of the design.  This is not currently implemented in
InstanceInfo, but will be soon.

This is done to make the behavior of Grand Central Views consistent with
the rest of the design/non-design interpretation and also to pave the way
to make them replaceable with layers.  (This makes them equivalent to
layers in the rationale.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>