[FIRRTL] Only add Seq Mem Ports in the design
Tighten the restrictions on when ports are added to memories.  Change this
to only occur for memories that are in the design.  Previously, this was
restricted to memories which were not under layers.  This makes the
behavior of this pass consistent across different non-design features and
this behavior can be easily extended simply by changing what the
InstanceInfo analysis deems is non-design.

As part of this, the entire logic for checking if a memory is legal to add
ports to can be collapsed to checking if it is in the effective design.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>