[circt-test] Add runner for circt-bmc (#7857)
Add a runner script for `circt-bmc` to circt-test. This allows CIRCT's
own model checker to be used to verify `verif.formal` operations in MLIR
inputs. The user currently has to pass the runner to circt-test
explicitly using the `-r` option. In the future, we'll want to add a
mechanism to circt-test to list available runners, check if the needed
software is installed on the system, and configure new local runners
through some config file.