[FIRRTL] Dedup: dedup private modules into public ones (#7877)
This change allows private modules to deduplicate in to public modules. It does this by swapping a private canonical module with a public module when found. Future public modules will not deduplicate into this group, as public modules can not dedup with other public modules in general. This also begins ignoring the convention of private modules, as it doesn't really carry any meaning. In the future we might want to disallow conventions to be specified on private modules.