[FIRRTL] Make IMDCE work for ops w/ regions/blocks
Fix a bug in FIRRTL's IMDCE Pass where it would not visit the blocks of
operations.  This can result in crashes for blocks which contain users of
FIRRTL modules, e.g., instances inside layerblocks of sv.ifdef.

This conceptually is two changes: (1) when marking a block live, the block
needs to be recursively walked and (2) when erasing ops, a recursive walk
is needed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>