[FIRRTL] Move LowerLayers (and pals) after IMDCE
Move `LowerLayers` after IMDCE.  Due to one small class of missed
optimizations, add a canonicalizer after `LowerLayers` to prevent Verilog
quality regressions (which were caught by an integration test).

This is working towards moving `LowerLayers` to the end of the FIRRTL
pipelinen so that inline layers can be used to crate `sv.ifdef` safely and
a feature can be added to force out of a layer block.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>