[FIRRTL] Remove all traces of OMIR JSON support.
This legacy system for representing an object model has been replaced
by FIRRTL classes, objects, and properties. This removes the old
system completely.

The main change is removing the EmitOMIR pass completely, but there
are several other things related to OMIR JSON to clean up. This
includes command line options and annotations for specifying input and
output OMIR JSON, as well as support for parsing the JSON and
scattering OMIR tracker annotations. Any tests related to the above
were also removed.

The only remaining legacy OMIR feature is the OMIRTracker annotation,
which is currently used by ResolvePaths and several passes. As we
rolled out the FIRRTL property support, we chose to re-use this
annotation for representing targets, so this is the one piece of the
old system that we have to keep for now.